refactoring on courses. working on user route
