+++ src/views/ClassroomRouter/services/reducer.js
@@ -1,9 +1,11 @@
 import { combineReducers } from 'redux'
 import dashboardReducer from '../views/Dashboard/services/reducer'
 import courseReducer from '../views/CourseRouter/services/reducer'
+import userReducer from '../views/User/services/reducer'
 
 // here we collect ALL reducers from views directory
 export default combineReducers({
 	dashboard: dashboardReducer,
-	course: courseReducer
+	course: courseReducer,
+	user: userReducer
 })
