+++ src/views/ClassroomRouter/views/CourseRouter/views/EditCourse/components/EditPanel/components/Section/index.js
@@ -1,0 +1,2 @@
+import Section from './Section'
+export default Section;
