+++ src/views/ClassroomRouter/views/CourseRouter/views/EditCourse/components/EditPanel/index.js
@@ -1,0 +1,2 @@
+import EditPanel from './EditPanel'
+export default EditPanel;
