+++ src/views/ClassroomRouter/views/User/User.js
@@ -1,13 +1,19 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux'
+import { getUser } from './services/actions'
 
 class User extends Component {
+
+	componentDidMount(){
+		this.props.getUser(this.props.match.params.userId);
+	}
+
 	render() {
 		// only show info, that user flagged as allowed to be shown
 		// show all possibilities for logged in user
 		return (
 			<div>
-				{JSON.stringify(this.props.authenticatedUser)}
+				{JSON.stringify(this.props.user)}
 			</div>
 		);
 	}
@@ -15,10 +21,17 @@ class User extends Component {
 
 let mapStateToProps = (state) => {
 	return {
-		...state.services
+		...state.views.classroom.user
+	}
+}
+
+let mapDispatchToProps = (dispatch) => {
+	return {
+		getUser: (userId) => dispatch(getUser(userId))
 	}
 }
 
 export default connect(
-	mapStateToProps
+	mapStateToProps,
+	mapDispatchToProps
 )(User)
