+++ src/views/ClassroomRouter/views/User/services/actionTypes.js
@@ -1,0 +1,3 @@
+export default {
+	API_GET_USER_BY_ID: 'API_GET_USER_BY_ID_USER'
+}
